commit post first debug round

+++ 0-preinstall.sh
@@ -32,22 +32,28 @@ echo -e "-Setting up US & Canada mirrors for faster downloads"
 echo -e "-------------------------------------------------------------------------"
 
 #download new mirrorlist and sort by fastest
-cd /etc/pacman.d/
-curl "https://archlinux.org/mirrorlist/?country=CA&country=US&protocol=http&protocol=https&ip_version=4&use_mirror_status=on" -o "mirrorlist.new"
-sed -i 's/^#Server/Server/' 'mirrorlist.new'
-rankmirrors -n 10 mirrorlist.new > mirrorlist
+if [[ ! -f /etc/pacman.d/mirrorlist.new ]]; then #ranking takes a second, will only already exist if we are running multiple times (during debugging)
+	cd /etc/pacman.d/
+	curl "https://archlinux.org/mirrorlist/?country=CA&country=US&protocol=http&protocol=https&ip_version=4&use_mirror_status=on" -o "mirrorlist.new"
+	sed -i 's/^#Server/Server/' 'mirrorlist.new'
+	echo "Ranking Mirrors"
+	rankmirrors -n 10 mirrorlist.new > mirrorlist
+	echo "Mirrors Ranked"
+else
+	echo "Mirrors already ranked"
+fi
 cd /
 
-mkdir /mnt
+mkdir -p /mnt
 
 
 echo -e "\nInstalling prereqs...\n$HR"
-pacman -S --noconfirm gptfdisk btrfs-progs
+pacman -S --noconfirm gptfdisk
 
 echo "--------------------------------------"
 echo -e "\nFormatting disk...\n$HR"
 echo "--------------------------------------"
-
+#read -p "continue?" con
 # disk prep
 sgdisk -Z ${DISK} # zap all on disk
 sgdisk -a 2048 -o ${DISK} # new gpt disk 2048 alignment
@@ -56,28 +62,28 @@ sgdisk -a 2048 -o ${DISK} # new gpt disk 2048 alignment
 sgdisk -n 1::+1M --typecode=1:ef02 --change-name=1:'BIOSBOOT' ${DISK} # partition 1 (BIOS Boot Partition)
 sgdisk -n 2::+100M --typecode=2:ef00 --change-name=2:'EFIBOOT' ${DISK} # partition 2 (UEFI Boot Partition)
 sgdisk -n 3::-10G --typecode=3:8300 --change-name=3:'ROOT' ${DISK} # partition 3 (Root), default start, minus 10G for swap
-sgdisk -n 4::- --typecode=4:8200 --change-name=4:'SWAP' ${Disk}
+sgdisk -n 4::- --typecode=4:8200 --change-name=4:'SWAP' ${DISK} #swap drive
 if [[ ! -d "/sys/firmware/efi" ]]; then
     sgdisk -A 1:set:2 ${DISK}
 fi
+
+#read -p "continue?" con
 
 # make filesystems
 echo -e "\nCreating Filesystems...\n$HR"
 if [[ ${DISK} =~ "nvme" ]]; then
 mkfs.vfat -F32 -n "EFIBOOT" "${DISK}p2"
-mkfs.ext4 -L "ROOT" "${DISK}p3" -f
+mkfs.ext4 -L "ROOT" "${DISK}p3" -F
 mount "${DISK}p3" /mnt
 mkswap "${DISK}p4"
 swapon "${DISK}p4"
 else
 mkfs.vfat -F32 -n "EFIBOOT" "${DISK}2"
-mkfs.ext4 -L "ROOT" "${DISK}3" -f
+mkfs.ext4 -L "ROOT" "${DISK}3" -F
 mount "${DISK}3" /mnt
 mkswap "${DISK}4"
 swapon "${DISK}4"
 fi
-;;
-esac
 
 mkdir /mnt/boot
 mount -t vfat -L EFIBOOT /mnt/boot/
@@ -91,6 +97,8 @@ if ! grep -qs '/mnt' /proc/mounts; then
     reboot now
 fi
 
+#read -p "ready to install. continue?" con
+
 echo "--------------------------------------"
 echo "-- Arch Install on Main Drive       --"
 echo "--------------------------------------"
@@ -100,7 +108,7 @@ echo "keyserver hkp://keyserver.ubuntu.com" >> /mnt/etc/pacman.d/gnupg/gpg.conf
 cp -R ${SCRIPT_DIR} /mnt/root/GraniteArch
 cp /etc/pacman.d/mirrorlist /mnt/etc/pacman.d/mirrorlist
 curl -L -O https://github.com/Michae11s/dots/archive/main.zip
-unzip main.zip
+unzip main.zip -q
 mv dots-main /mnt/dots/
 
 echo "--------------------------------------"
@@ -112,3 +120,4 @@ fi
 echo "--------------------------------------"
 echo "--   SYSTEM READY FOR 1-setup       --"
 echo "--------------------------------------"
+#read -p "continue?" con
